Исправления по результатм проверки. Установка корректных имен контроллерам. Модификация web.xml.
