+++ auth/src/main/java/controllers/admin/RegistrationUrlController.java
@@ -1,4 +1,4 @@
-package servlets.admin;
+package controllers.admin;
 
 import classes.Role;
 import org.apache.log4j.Logger;
@@ -9,9 +9,6 @@ import org.springframework.web.bind.annotation.RequestParam;
 import services.RegistrationUrlService;
 import services.impl.RegistrationUrlServiceImpl;
 
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 @Controller
