+++ auth/src/main/java/controllers/ajax/RemoveCityById.java
