--- auth/src/main/java/servlets/filters/CharacterEncodingFilter.java
@@ -1,22 +1,0 @@
-package servlets.filters;
-
-import javax.servlet.*;
-import java.io.IOException;
-
-public class CharacterEncodingFilter implements Filter {
-    @Override
-    public void init(FilterConfig filterConfig) throws ServletException {
-
-    }
-
-    @Override
-    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
-        request.setCharacterEncoding("UTF-8");
-        chain.doFilter(request, response);
-    }
-
-    @Override
-    public void destroy() {
-
-    }
-}
