+++ auth/src/main/webapp/WEB-INF/web.xml
@@ -10,7 +10,7 @@
     <!--ENCODING FILTER-->
     <filter>
         <filter-name>encodingFilter</filter-name>
-        <filter-class>servlets.filters.CharacterEncodingFilter</filter-class>
+        <filter-class>controllers.filters.CharacterEncodingFilter</filter-class>
     </filter>
 
     <filter-mapping>
@@ -23,7 +23,7 @@
 
     <filter>
         <filter-name>AuthorizationFilter</filter-name>
-        <filter-class>servlets.filters.AuthorizationFilter</filter-class>
+        <filter-class>controllers.filters.AuthorizationFilter</filter-class>
     </filter>
 
     <filter-mapping>
@@ -63,7 +63,7 @@
     </servlet-mapping>
     <context-param>
         <param-name>contextConfigLocation</param-name>
-        <param-value>/WEB-INF/DispatcherServlet-servlet.xml</param-value>
+        <param-value>/WEB-INF/*-servlet.xml</param-value>
     </context-param>
 
     <listener>
